+++ packages/ui-library/CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 ## [1.8.7-beta.2](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.8.7-beta.0...@ab.uitools/ui-library@1.8.7-beta.2) (2026-04-20)
+## [1.8.6](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.8.5...@ab.uitools/ui-library@1.8.6) (2026-04-20)
+
+### Bug Fixes
+
+- remove additional classes for custom card-select ([9c1cd0b](https://github.com/ab-ui-tools/ab-ui-tools/commit/9c1cd0b57ebddfd736801bfcdb52120682515242))
+
+## [1.8.5](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.8.4...@ab.uitools/ui-library@1.8.5) (2026-04-20)
+
+### Bug Fixes
+
+- table column settings issue ([584a097](https://github.com/ab-ui-tools/ab-ui-tools/commit/584a09730c2f6c383a5f373a36ba22e068438701))
+
+## [1.8.4](https://github.com/ab-ui-tools/ab-ui-tools/compare/@ab.uitools/ui-library@1.8.4-beta.1...@ab.uitools/ui-library@1.8.4) (2026-04-18)
 
 ### Bug Fixes